stm32 CORDIC: split into multiple files
